Required changes to rename Permission Tree New component
