+++ .storybook/manager-head.html
@@ -7,7 +7,7 @@
     #components-forms-question,    
     #components-hierarchy, 
     #components-order-details,
-    #components-permission-tree,
+    #components-permission-tree-new,
     #components-questions,    
     #components-syntax-highlighter,
     #components-icon-list,
