+++ raaghu-components/src/rds-comp-permission-tree/rds-comp-permission-tree.stories.tsx
@@ -1,92 +1,99 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import RdsCompPermissionTree from "./rds-comp-permission-tree";
+import TreeNode from './rds-comp-permission-tree';
 
 
 const meta: Meta = { 
     title: "Components/Permission Tree",
-    component: RdsCompPermissionTree,
+    component: TreeNode,
     parameters: {
         layout: 'padded',
     },
     tags: ['autodocs'],
     argTypes: {
     },
-} satisfies Meta<typeof RdsCompPermissionTree>;
+} satisfies Meta<typeof TreeNode>;
 
 export default meta;
-type Story = StoryObj<typeof RdsCompPermissionTree>;
+type Story = StoryObj<typeof TreeNode>;
 
 export const Default: Story = {
     args: {
-        permissions: [
+        treeData: [
             {
-                displayName: "[Test edition scope feature]",
-                id: "testEditionScopeFeature",
-                isSelected: false,
-                isIntermediate: false,
-                disabled: false,
-                parent_id: "",
-                children: []
-            },
-            {
-                displayName: "Chat",
-                id: "chat",
-                parent_id: "",
-                isSelected: false,
-                isIntermediate: false,
-                disabled: false,
+                data: { id: 1 },
+                label: 'Root',
+                selected: false,
+                level: 0,
                 children: [
                     {
-                        displayName: "Chat with host",
-                        id: "chatwithhost",
-                        parent_id: "chat",
-                        isSelected: false,
-                        isIntermediate: false,
-                        disabled: false,
-                        children: []
+                        data: { id: 2 },
+                        label: 'Child 1',
+                        selected: false,
+                        level: 1,
+                        children: [
+                            {
+                                data: { id: 3 },
+                                label: 'Grandchild 1',
+                                selected: false,
+                                level: 2,
+                                children: [],
+                            },
+                            {
+                                data: { id: 4 },
+                                label: 'Grandchild 2',
+                                selected: false,
+                                level: 2,
+                                children: [],
+                            },
+                        ],
                     },
                     {
-                        displayName: "Chat with other tenants",
-                        id: "chatwithothertenats",
-                        parent_id: "chat",
-                        isSelected: false,
-                        isIntermediate: false,
-                        disabled: false,
-                        children: []
+                        data: { id: 5 },
+                        label: 'Child 2',
+                        selected: false,
+                        level: 1,
+                        children: [],
                     },
                 ],
             },
             {
-                displayName: "Maximum user count",
-                id: "maximumUserCount",
-                parent_id: "",
-                isSelected: false,
-                isIntermediate: false,
-                disabled: false,
-                children: []
-            },
-            {
-                displayName: "Test check feature",
-                id: "testCheckFeature",
-                parent_id: "",
-                isSelected: false,
-                isIntermediate: false,
-                disabled: false,
-                children: []
-            },
-            {
-                displayName: "Test check feature",
-                id: "testCheckFeature",
-                parent_id: "",
-                isSelected: true,
-                isIntermediate: false,
-                disabled: false,
-                children: []
+                data: { id: 2 },
+                label: 'New',
+                selected: false,
+                level: 1,
+                children: [
+                    {
+                        data: { id: 2 },
+                        label: 'Child 1',
+                        selected: false,
+                        level: 1,
+                        children: [
+                            {
+                                data: { id: 3 },
+                                label: 'Grandchild 1',
+                                selected: false,
+                                level: 2,
+                                children: [],
+                            },
+                            {
+                                data: { id: 4 },
+                                label: 'Grandchild 2',
+                                selected: false,
+                                level: 2,
+                                children: [],
+                            },
+                        ],
+                    },
+                    {
+                        data: { id: 5 },
+                        label: 'Child 2',
+                        selected: false,
+                        level: 1,
+                        children: [],
+                    },
+                ],
             },
         ],
-    
-        selectedPermissions: () => {
-        }
     }
 } satisfies Story;
